# static/openapi.yaml
openapi: 3.0.3
info:
  title: Api-SportMatch
  description: >
    API REST de SportMatch. Fase 2: autenticación, usuarios y administradores.
    Fase 3 (en construcción): deportes, cuestionario de registro, eventos,
    inscripciones y valoraciones, organizadores y validación, administración
    y soporte. Ver /docs/entidades_completas.md en el repositorio para el
    catálogo completo de las 89 entidades del modelo.
  version: 0.5.0
servers:
  - url: http://localhost:5000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /api/organizadores:
    get:
      summary: Listar organizadores (público; solo aprobados. ?estado_validacion= para admin)
      tags: [Organizadores]
      security: []
      responses:
        "200": { description: Lista de organizadores }
    post:
      summary: Crear mi perfil de organizador (autenticado; uno por usuario, queda "pendiente")
      tags: [Organizadores]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre_comercial]
              properties:
                nombre_comercial: { type: string }
                descripcion: { type: string }
                correo_contacto: { type: string }
                telefono_contacto: { type: string }
                ciudad_id: { type: integer }
      responses:
        "201": { description: Organizador creado }
        "409": { description: Ya tienes un perfil de organizador }
  /api/organizadores/{organizador_id}:
    get:
      summary: Detalle de un organizador
      tags: [Organizadores]
      security: []
      parameters:
        - { name: organizador_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Organizador }
    put:
      summary: Editar mi organizador (miembro propietario/administrador, o admin)
      tags: [Organizadores]
      parameters:
        - { name: organizador_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Organizador actualizado }
  /api/organizadores/{organizador_id}/miembros:
    post:
      summary: Agregar un miembro al equipo del organizador
      tags: [Organizadores]
      parameters:
        - { name: organizador_id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [usuario_id]
              properties:
                usuario_id: { type: integer }
                rol: { type: string, enum: [propietario, administrador, editor] }
      responses:
        "201": { description: Miembro agregado }
  /api/organizadores/{organizador_id}/solicitudes:
    post:
      summary: Enviar el organizador a revisión, opcionalmente con documentos anexos
      tags: [Organizadores]
      parameters:
        - { name: organizador_id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                documentos:
                  type: array
                  items:
                    type: object
                    properties:
                      tipo_documento: { type: string, example: "ine" }
                      numero_documento: { type: string }
      responses:
        "201": { description: Solicitud enviada (organizador pasa a "en_revision") }
        "409": { description: Ya hay una solicitud pendiente }
...
  /api/organizadores/solicitudes/{solicitud_id}/revisar:
    put:
      summary: Aprobar o rechazar una solicitud de validación (solo admin)
      tags: [Organizadores]
      parameters:
        - { name: solicitud_id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [aprobar, rechazar] }
                comentarios: { type: string }
                motivo_rechazo: { type: string }
      responses:
        "200": { description: Organizador y solicitud actualizados }
...